Consolidate PCP kernel tests into rpa_v3_cp test file
The PCP kernel tests added on this branch duplicated the file that landed
in main via #3166 (tests/kernels/rpa_v3_cp/ragged_paged_attention_kernel_cp_test.py).
Fold the two new additions -- the _pi2 page-index helper and
test_pcp_fused_current_phase_write -- into that file and drop the
duplicate tests/kernels/ragged_paged_attention_pcp_test.py.
